Refactor in a geometry package
