+++ diagram-util/src/main/java/com/powsybl/diagram/util/forcelayout/geometry/Vector2D.java
@@ -1,10 +1,11 @@
-/**
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
+/*
+ * Copyright (c) 2021-2025, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.diagram.util.forcelayout;
+package com.powsybl.diagram.util.forcelayout.geometry;
 
 /**
  * @author Mathilde Grapin {@literal <mathilde.grapin at rte-france.com>}
