+++ network-area-diagram/src/main/java/com/powsybl/nad/layout/BasicForceLayout.java
@@ -8,7 +8,7 @@
 package com.powsybl.nad.layout;
 
 import com.powsybl.diagram.util.forcelayout.ForceLayout;
-import com.powsybl.diagram.util.forcelayout.Vector2D;
+import com.powsybl.diagram.util.forcelayout.geometry.Vector2D;
 import com.powsybl.nad.model.Edge;
 import com.powsybl.nad.model.Graph;
 import com.powsybl.nad.model.Node;
@@ -72,12 +72,12 @@ public class BasicForceLayout extends AbstractLayout {
     }
 
     private void setInitialPositions(ForceLayout<Node, Edge> forceLayout, Graph graph) {
-        Map<Node, com.powsybl.diagram.util.forcelayout.Point> initialPoints = getInitialNodePositions().entrySet().stream()
+        Map<Node, com.powsybl.diagram.util.forcelayout.geometry.Point> initialPoints = getInitialNodePositions().entrySet().stream()
                 // Only accept positions for nodes in the graph
                 .filter(nodePosition -> graph.getNode(nodePosition.getKey()).isPresent())
                 .collect(Collectors.toMap(
                     nodePosition -> graph.getNode(nodePosition.getKey()).orElseThrow(),
-                    nodePosition -> new com.powsybl.diagram.util.forcelayout.Point(
+                    nodePosition -> new com.powsybl.diagram.util.forcelayout.geometry.Point(
                             nodePosition.getValue().getX() / SCALE,
                             nodePosition.getValue().getY() / SCALE)
                 ));
